% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{compile_sim_ellip}
\alias{compile_sim_ellip}
\title{Setup Underlying Simulation by Compiling It}
\usage{
compile_sim_ellip(showCompilerOutput = FALSE)
}
\arguments{
\item{showCompilerOutput}{If TRUE, show output from C++ compiler; may be useful for 
debugging compiler issues}
}
\value{
None
}
\description{
Compiles Nimble function describing underlying trait simulation into C++ code
}
