% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_niche_fill_sim.R
\name{continue_sim_radiation}
\alias{continue_sim_radiation}
\title{Continue Niche Filling Simulation Function}
\usage{
continue_sim_radiation(sim_ob, time_steps = 10000,
  arrest_speciation = FALSE)
}
\arguments{
\item{sim_ob}{A simulation object created by a previous run of \code{\link{sim_radiation}}}

\item{time_steps}{Number of additional time steps to run the simulation.}

\item{arrest_speciation}{Logical. TRUE if the speciation part of the simulation should be 
arrested. This is usually done to allow the simulation to come to some equilibrium state.}
}
\value{
A niche_fill_sim object containing the final simulation object,
a set of intermediate phylogenies, and the parameters used to run the simulation
}
\description{
Function to continue running the niche filling simulation on a simulation that has already been run,
picking up where it left off.
}
