% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_gen.R
\name{example_parms}
\alias{example_parms}
\title{Function to generate example paraeter objects.}
\usage{
example_parms(name = c("simple_peaks_2", "blackhole_comp"))
}
\description{
Function to generate example paraeter objects.
}
