% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_results.R
\name{extract_final_SR}
\alias{extract_final_SR}
\title{Extract Final Species Richness of Simulation}
\usage{
extract_final_SR(sim_ob)
}
\arguments{
\item{sim_ob}{Simulation object from which to extract the final species richness}
}
\value{
The final species richness value in the simulation
}
\description{
Function to extract the final species richness in a simulation
}
