% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_results.R
\name{extract_params}
\alias{extract_params}
\title{Extract Useful Parameters of Simulation Object}
\usage{
extract_params(sim_ob)
}
\arguments{
\item{sim_ob}{Simulation object from which to extract the final species richness}
}
\value{
Single row data.frame of parameter values
}
\description{
Function to extract the useful parameters used in the simulation
}
