% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_2_helpers.R
\name{generate_all_params}
\alias{generate_all_params}
\title{Generate Hopeful Niche Monster parameters}
\usage{
generate_all_params(potent_vol = 1, total_vol = 1, d = 2L, mult = 0.1,
  gamma_i = 1, V_gi = 0.01, b_rate = 0.01, total_time = 50000,
  D = 1.25)
}
\arguments{
\item{param_df_list}{A data.frame with list columns containing paramter values to
be run in the simulation.}

\item{parallel}{Logical: Should the simulation be run in parallel?}

\item{cores}{How many cores to use?}

\item{save_folder}{Folder to save simulation results to.}

\item{num_tries}{Number of times to try any simulation if their are errors before
giving up.}

\item{prefix}{Prefix to prepend to filenames.}

\item{compress}{Compression method to use: "none", "gz" ,"bz", or "xz".}
}
\value{
A vector of filenames pointing to where each simulation object is saved
}
\description{
Need to fill this in ... ignore the below, it is wrong.
}
