% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_gen.R
\name{generate_landscape}
\alias{generate_landscape}
\title{Generate a Random Carrying Capacity Landscape}
\usage{
generate_landscape(potent_vol = 1, total_vol = 1, num_peaks = 6,
  h_to_sig_ratio = 2, P_min_max = c(0.9, 1.1), dirichlet_param = 1.5,
  d = 2, P = 8, a_prop = 0.01)
}
\arguments{
\item{potent_vol}{Total potential volume of the entire landscape.}

\item{total_vol}{Total approximate volume of the entire landscape.}

\item{num_peaks}{Number of fitness peaks to place on the landscape.}

\item{h_to_sig_ratio}{Ratio of the height of each peak to its width (or sigma value).}

\item{P_min_max}{Vector of length 2 giving the minimum and maximum peak super-gaussian
parameters to use.}

\item{dirichlet_param}{The alpha parameter of the dircihlet distribution for 
determining how the volume should be split between the peaks.}

\item{d}{The number of dimensions of the landscape}

\item{P}{Super-gaussian parameter for whole landscape multiplier}

\item{a_prop}{Minimum carrying capacity within the landscape as a proportion of the maximum
height}
}
\value{
List of parameter values that can be used in the carrying capacity function
to generate the landscape
}
\description{
Function to generate a random carrying capacity landscape obeying certain constraints. 
It distributes peaks uniformly within an N-ball, which is approximated by the overall
Super-gaussian multiplier distribution. Peak heights and widths are chosen to
satisfy a particular ration and to add up to the total desired niche 'volume'.
}
