% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_pop_sim.R
\name{generate_landscape_old}
\alias{generate_landscape_old}
\title{Generate Random Carrying Capacity Landscape}
\usage{
generate_landscape_old(u = 2, h_mean = 1, h_var = 0.5,
  bi_mean = c(0, 0), bi_var = c(3, 3), h0 = 1, sig0i = c(5, 5),
  P0i = c(1.5, 1.5), sigi_mean = c(1, 1), sigi_var = c(1, 1),
  Pi_mean = c(1, 1), Pi_var = c(0.25, 0.25), a = 0.5,
  Piz_min = 0.25, hz_min = 0.25)
}
\arguments{
\item{u}{Number of carrying capacity peaks to generate}

\item{h_mean}{The mean height of the carrying capacity peaks}

\item{h_var}{The variance in height of the peaks}

\item{bi_mean}{The mean peak centre value for each of d dimensions; vector of length d}

\item{bi_var}{The variance in peak centre values for each of d dimensions; vector of length d}

\item{h0}{The maximum height of the carrying capacity on the landscape}

\item{sig0i}{The width of the carrying capacity landscape, e.g. how far it extends before dropping off
in each of d dimensions; vector of length d}

\item{P0i}{Total super-gaussian parameter; determines how quickly or gradually the carrying
capacity drops off near the landscape borders in each dimension. Higher values give more extreme drop-offs;
vector of length d}

\item{sigi_mean}{The mean width of peaks for each of d dimensions; vector of length d}

\item{sigi_var}{The variance in peak widths for each of d dimensions; vector of length d}

\item{Pi_mean}{The mean super-gaussian parameter of peaks for each of d dimensions; determines
the sharpness of peak drop-off; vector of length d}

\item{a}{Minimum values of carrying capacity within landscape limits}

\item{hz_min}{Minimum height allowed for any peak}
}
\value{
List of parameter values that can be passed on to other functions such as 
\code{\link{K_func}}
}
\description{
Function to generate a random carrying capacity landscape as a mix of super-gaussians
}
