% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_2_helpers.R
\name{generate_random_params_2}
\alias{generate_random_params_2}
\title{Generate Hopeful Niche Monster random parameters}
\usage{
generate_random_params_2(nreps = 10, total_vol_range = c(0.05, 1),
  mult_range = c(0.01, 1), gamma_range = c(0.01, 1), V_gi_range = c(0.001,
  0.1), d_range = c(1L, 4L), b_rate_range = c(1e-04, 0.01),
  tot_time_range = c(20000, 1e+05))
}
\arguments{
\item{param_df_list}{A data.frame with list columns containing paramter values to
be run in the simulation.}

\item{parallel}{Logical: Should the simulation be run in parallel?}

\item{cores}{How many cores to use?}

\item{save_folder}{Folder to save simulation results to.}

\item{num_tries}{Number of times to try any simulation if their are errors before
giving up.}

\item{prefix}{Prefix to prepend to filenames.}

\item{compress}{Compression method to use: "none", "gz" ,"bz", or "xz".}
}
\value{
A vector of filenames pointing to where each simulation object is saved
}
\description{
Need to fill this in ... ignore the below, it is wrong.
}
