% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_gen.R
\name{generate_random_params_QMC}
\alias{generate_random_params_QMC}
\title{Generate Quasi Monte Carlo Random Parameters for Simulation}
\usage{
generate_random_params_QMC(reps = 10000, potent_vol_range = c(1, 20),
  total_vol_range = c(5, 5), num_peaks_range = c(2L, 10L),
  P_min_range = c(0.9, 0.9), P_max_range = c(1.1, 1.1),
  prop_variance_range = c(0.1, 0.5), d_range = c(2L, 2L), P_range = c(1,
  1), e_var_range = c(0.1, 0.4), gamma_range = c(0.01, 1),
  b_rate_range = c(0.001, 0.003), D_range = c(1, 1),
  h_to_sig_ratio_range = c(0.5, 2), a_prop_range = c(0.05, 0.05),
  tot_time_range = c(50000, 50000), V_range = c(0.01, 0.01),
  mult_range = c(1, 1), C_range = c(1, 1), c_var_range = c(0.1, 0.1))
}
\arguments{
\item{reps}{Number of random sets of paramters to generate}

\item{xx_range}{Vectors of length two, specifying the minimum and maximum values of
parameter \code{xx} in the model}
}
\value{
A data.frame containing the random parameter values
}
\description{
Function that takes a set of minimum and maximum values and generates uniform random values
for each parameter within a hypercube using QuasiMC
}
