% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_gen.R
\name{integrate_fun}
\alias{integrate_fun}
\title{Numerically Integrate Carrying Capacity Landscape}
\usage{
integrate_fun(K_func, lower, upper, ...)
}
\arguments{
\item{K_func}{R function specifying the carrying capacity function.}

\item{lower}{Lower bound to integrate across; vector of length d, where d is the number of
dimensions}

\item{upper}{Upper bound to integrate across; vector of length d.}

\item{...}{Named arguments to be passed to fun.}
}
\value{
The total volume of the landscape within the lower and upper bounds.
}
\description{
Integrate a carrying capacity landscape to estimate its 'volume'
}
