% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_pop_sim.R
\name{K_func}
\alias{K_func}
\title{Carrying Capacity Landscape Function (R version)}
\usage{
K_func(xi, h0 = 1, sig0i = c(5, 5), P0i = c(1.5, 1.5), hz = c(1,
  1), biz = matrix(c(-1, -1, 1, 1), nrow = 2, ncol = 2),
  sigiz = matrix(c(1, 1, 1, 1), nrow = 2, ncol = 2), Piz = matrix(c(1,
  1, 1, 1), nrow = 2, ncol = 2), a = 0.1, use_ellip = TRUE)
}
\arguments{
\item{xi}{X values to evaluate under carrying capacity function; a vector of length d, 
where d is the number of dimensions of the landscape}

\item{h0}{Total maximum height of the carrying capacity landscape}

\item{sig0i}{Total width of landscape in all dimensions; determines how far from zero the carrying capacity
drops off to nearly zero; vector of length d}

\item{P0i}{Total super-gaussian parameter; determines how quickly or gradually the carrying
capacity drops off near the landscape borders in each dimension. Higher values give more extreme drop-offs;
vector of length d}

\item{hz}{Maximum height of each of u peaks in the landscape; vector of length u}

\item{biz}{Centre of each of u peaks in the landscape for each of d dimensions; 
matrix of dimension d by u}

\item{sigiz}{Width of each of u peaks for each of d dimensions; matrix of length d by u}

\item{Piz}{Super-gaussian parameter for each of u peaks for each of d dimensions; 
matrix of dimension d by u}

\item{a}{Minimum values of carrying capacity within landscape limits}
}
\value{
Value of carrying capacity at coordinates \code{xi}
}
\description{
Carrying capacity function written in R, mostly for visualization or testing purposes.
Returns the value of the carrying capacity landscape at coordinate \code{xi}.
}
