% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{make_df_from_sim}
\alias{make_df_from_sim}
\title{Function to extract trait history from simulation object as a data_frame}
\usage{
make_df_from_sim(sim_ob)
}
\arguments{
\item{sim_ob}{nichefillr_sim object containing the results of a simulation.}
}
\description{
Function to extract trait history from simulation object as a data_frame
}
