% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_results.R
\name{make_ode_df}
\alias{make_ode_df}
\title{Convert ODE Output into a Data.frame}
\usage{
make_ode_df(ode_mat, d, m, thin = 100)
}
\arguments{
\item{ode_mat}{An ode object matrix}

\item{d}{The number of trait dimensions in the simulation}

\item{m}{The number of species in the simulation}

\item{thin}{How much to thin the ode output, returns 1 out of every \code{thin} timesteps}
}
\value{
A data.frame with simulation output
}
\description{
Function to convert the ode objects outputted by \code{deSolve} into a data.frame
}
