% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_1_helpers.R
\name{make_params_lists}
\alias{make_params_lists}
\title{Create Lists of Parameters from a Parameter Dataframe}
\usage{
make_params_lists(param_df, trait_hist_prop = 0.3)
}
\arguments{
\item{param_df}{Dataframe containing parameters to be used to generate random
carrying capacity landscapes and for competition and macroevolutionary parts
of the simulation}
}
\value{
A data.frame containing the final parameter lists in a list column. This can be 
used by \code{\link{sim_radiation}}
}
\description{
Function to generate parameter lists in list columns using a data.frame with
parameter values used to generate random landscapes
}
\details{
Note: this function can take some time as it numerically integrates each landscape
to make sure the integral adds up to the correct total volume
}
