% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_results.R
\name{make_sim_df}
\alias{make_sim_df}
\title{Extract Data.frame of Simulation Data}
\usage{
make_sim_df(full_sim_ob, thin_to = 10000, include_trees = TRUE,
  summarise_trees = FALSE, PD = "PSE")
}
\arguments{
\item{full_sim_ob}{A simulation object as produced by \code{\link{sim_radiation}}}

\item{thin_to}{How many timesteps of data to return. The function will thin the output to match
this length.}
}
\value{
A data.frame with simulation data for \code{thin_to} timesteps
}
\description{
Function to extract useful data from a simulation object and return it in a data.frame
}
