% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_sim.R
\name{make_tree_ob}
\alias{make_tree_ob}
\title{Initialize Tree Object for Macroevolution Simulation}
\usage{
make_tree_ob(b_rate = 0.001, init_traits, e_var, init_Ns, trait_hist = TRUE,
  save_tree = TRUE)
}
\arguments{
\item{b_rate}{Pure birth rate for new species, here just to get first branching;}

\item{init_traits}{Vector of initial trait values of length n_trait, where n_trait is
the number of traits to be simulated}

\item{e_var}{Vector of gaussian variances determining rate of evolution along branches; of 
length n_trait; This is the Brownian motion rate parameter}
}
\value{
Initialized object with macroevolutionary simulation setup
}
\description{
Initializes an object used by the macroevolutionary simulation to store data
 about species, their relationships, and their population sizes and trait values
}
