% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot.nichfillr_sim}
\alias{plot.nichfillr_sim}
\title{Plot a \code{nichefillr_sim} object. If the simulation object contains a trait history,
this function will by default plot a 'pollock plot'. If the object contains no trait history it
will plot the trait values of the extant species, extinct species (optionally), 
and ancestors (optionally). Currently only work for 2 dimensional simulations, eventually
will work with higher dimensions through t-sne dimension reduction.}
\usage{
\method{plot}{nichfillr_sim}(x, fitness_contour = TRUE, contour_res = 100,
  expand_factor = 0.1)
}
\description{
Plot a \code{nichefillr_sim} object. If the simulation object contains a trait history,
this function will by default plot a 'pollock plot'. If the object contains no trait history it
will plot the trait values of the extant species, extinct species (optionally), 
and ancestors (optionally). Currently only work for 2 dimensional simulations, eventually
will work with higher dimensions through t-sne dimension reduction.
}
