% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_1_helpers.R
\name{run_sim_and_save}
\alias{run_sim_and_save}
\title{Helper Function to Run Simulation and Save It}
\usage{
run_sim_and_save(parms, file_name, num_tries, compress = "gz")
}
\arguments{
\item{parms}{Simulation Parameters.}

\item{file_name}{Filename to save simulation results to.}

\item{num_tries}{Number of times to try each simulation in the case of errors before
giving up and moving on.}

\item{compress}{Compression method to use: "none", "gz" ,"bz", or "xz".}
}
\value{
A data.frame containing the filename saved to.
}
\description{
Helper Function to Run Simulation and Save It
}
