% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_2_helpers.R
\name{run_sims_2}
\alias{run_sims_2}
\title{Run Simulations from a Parameter List Data.frame for Hopeful Niche Monster Paper}
\usage{
run_sims_2(param_df, parallel = TRUE, cores = 8, save_folder = "results",
  num_tries = 10, prefix = "", compress = "gz")
}
\arguments{
\item{parallel}{Logical: Should the simulation be run in parallel?}

\item{cores}{How many cores to use?}

\item{save_folder}{Folder to save simulation results to.}

\item{num_tries}{Number of times to try any simulation if their are errors before
giving up.}

\item{prefix}{Prefix to prepend to filenames.}

\item{compress}{Compression method to use: "none", "gz" ,"bz", or "xz".}

\item{param_df_list}{A data.frame with list columns containing paramter values to
be run in the simulation.}
}
\value{
A vector of filenames pointing to where each simulation object is saved
}
\description{
Function to take a parameter data.frame with list columns and setup
and run the simulation in parallel (using \code{multidplyr})
}
