% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_1_helpers.R
\name{run_sims_paper_1}
\alias{run_sims_paper_1}
\title{Run Simulations for Paper #1}
\usage{
run_sims_paper_1(num_reps, num_cores, folder = "results", prefix,
  compress = "gz", param_df_list = NULL, ...)
}
\arguments{
\item{num_reps}{Number of simulation reps to run.}

\item{num_cores}{Number of cores to use in parallel computations.}

\item{folder}{Name of folder to place output files.}

\item{prefix}{Prefix of type character to prepend to any filenames.}

\item{...}{Other parameters passed onto \code{\link{generate_random_params}}}
}
\value{
None
}
\description{
Function to run simulations used in Paper #1
}
