% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_sims.R
\name{sim_radiation_multi}
\alias{sim_radiation_multi}
\title{Function that takes a list of parameter objects and runs them all, with optional parallelization}
\usage{
sim_radiation_multi(parm_list, save_tree = TRUE, progress = TRUE,
  trait_hist = TRUE, trait_hist_prop = 0.01, ncpus = NULL,
  save_folder = NULL, save_prefix = "sim_", compress = "gz")
}
\description{
Function that takes a list of parameter objects and runs them all, with optional parallelization
}
