% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_pop_sim.R
\name{trait_pop_sim_de}
\alias{trait_pop_sim_de}
\title{Calculate Niche Filling Trait and Population Simulation Derivative}
\usage{
trait_pop_sim_de(t, y, parms)
}
\arguments{
\item{t}{Time step of the trait evolution simulation; this is for compatability with 
\code{deSolve} and is ignored}

\item{y}{Current value of the traits and populations}

\item{parms}{Named list containing all necessary parameters}
}
\value{
List of derivatives compatible with \code{deSolve}
}
\description{
Calculates the derivative of all traits in the niche-filling simulation
For use with \code{deSolve} functions to run the trait evolution simulation. User probably 
won't have reason to use this directly
}
