% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_sim.R
\name{update_br_lens}
\alias{update_br_lens}
\title{Function to Update Branch-lengths (and Traits) in Macroevolutionary Simulation}
\usage{
update_br_lens(tree_ob, add_br_len, e_var = NULL)
}
\arguments{
\item{tree_ob}{The simulation object to update}

\item{add_br_len}{How much branch-length should be added (in units of simulation time)}

\item{e_var}{Vector of evolutionary rates of length n_trait, where n_trait is the number
of traits being simulated}
}
\value{
An updated simulation object
}
\description{
Updates branch-length by adding to time to every branch; also updates trait values using
Brownian motion simulation of the same time period. This is used at the beginning of
the simulation to setup the first split of two species, and to grow branches within
the main loop (with no trait evolution).
}
