% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_sim.R
\name{update_tree_ob_extinction}
\alias{update_tree_ob_extinction}
\title{Update Simulation Object with Extinction Event}
\usage{
update_tree_ob_extinction(tree_ob, extinct_tip)
}
\arguments{
\item{tree_ob}{The simulation object to update}

\item{extinct_tip}{Integer expressing which tip number to go extinct (stop updating)}
}
\value{
An updated simulation object
}
\description{
Function to update simulation object by removing a species from the simulation
}
