% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_sim.R
\name{update_tree_ob_speciate}
\alias{update_tree_ob_speciate}
\title{Update Simulation Object with Speciation Event}
\usage{
update_tree_ob_speciate(tree_ob, split_tip, mult = 1)
}
\arguments{
\item{tree_ob}{The simulation object to update}

\item{split_tip}{Integer expressing which tip number to split into two sister species}

\item{mult}{Multiplier which determines the proportional increase in trait evolution
rate at a speciation event. This causes species to diverge more rapidly than usual if
greater than one. Generally used to make sure a speciation is accompanied by a 'jump'
in trait values, which helps them not immediately go extinct from competition}
}
\value{
An updated simulation object
}
\description{
Function to update simulation object by adding add_br_len to the tips, 
then splitting a tip into two species
}
