% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_gen.R
\name{vis_K_landscape}
\alias{vis_K_landscape}
\title{Visualize a Carrying Capacity Landscape using RGL}
\usage{
vis_K_landscape(K_parms, total_rad)
}
\arguments{
\item{K_parms}{A names list of parameters describing the landscape; currently only
excepts two dimensional landscapes}
}
\value{
None
}
\description{
This function uses the R package \code{rgl} to visualize a carrying capacity lanscape 
function. Currently only works for two-dimensional landscapes.
}
