% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_gen.R
\name{example_parms}
\alias{example_parms}
\title{Function to generate example parameter objects.}
\usage{
example_parms(name = c("simple_peaks_2", "blackhole_comp"))
}
\arguments{
\item{name}{Name of example parameter set. See details for list of available options (not done yet).}
}
\value{
A list containing a set of simulation parameters.
}
\description{
Function to generate example parameter objects.
}
