% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{sim_animation}
\alias{sim_animation}
\title{Function to animate the results of a simulation. Currently only supports simulations
done in two dimensions. Will eventually support higher dimensions using t-sne dimension
reduction.}
\usage{
sim_animation(sim_ob, file_name = NULL, view = FALSE, expand_factor = 0.1,
  contour_res = 25, height = 300, width = 400, res = 72)
}
\arguments{
\item{sim_ob}{nichefillr_sim object containing the results of a simulation.}

\item{file_name}{(optional) Path to to save the animation to. If NULL, and
view = TRUE, animation will be output to the rstudio viewer.}

\item{expand_factor}{Factor by which to expand x and y axes beyond the minimum and
maximum found in the simulated species (allows more of the fitness landscape context to ve
shown)}

\item{contour_res}{Resolution for the fitness contours, in terms of how many points per
axis to calculate fitness for.}
}
\description{
Function to animate the results of a simulation. Currently only supports simulations
done in two dimensions. Will eventually support higher dimensions using t-sne dimension
reduction.
}
